% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxillary_functions.R
\name{analyzeTips}
\alias{analyzeTips}
\title{Determines which tips are missing from trees, and how often.}
\usage{
analyzeTips(multiPhyloObject)
}
\arguments{
\item{multiPhyloObject}{The tree object.}
}
\description{
Determines which tips are missing from trees, and how often.
}
