% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_trees.R
\name{bootStrapWithNoiseFromNorm}
\alias{bootStrapWithNoiseFromNorm}
\title{noise}
\usage{
bootStrapWithNoiseFromNorm(originalMatrix, standardDev, numTrees = 1000,
  rootTree = NULL)
}
\arguments{
\item{originalMatrix}{hkjh}

\item{standardDev}{kjhjkh}

\item{numTrees}{hkjh}

\item{rootTree}{kjhkj}
}
\description{
noise
}
