% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_trees.R
\name{bootStrapWithNoiseOverRange}
\alias{bootStrapWithNoiseOverRange}
\title{Bootstraps}
\usage{
bootStrapWithNoiseOverRange(originalMatrix, rangeMin = 1, rangeMax = 1000,
  numTrees = 1000, rootTree = NULL)
}
\arguments{
\item{originalMatrix}{hkjh}

\item{rangeMin}{hkj}

\item{rangeMax}{khj}

\item{numTrees}{khj}

\item{rootTree}{hkj}
}
\description{
Bootstraps
}
