% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_trees.R
\name{generateBootstrapTrees}
\alias{generateBootstrapTrees}
\title{Bootstraps}
\usage{
generateBootstrapTrees(distanceMatrix, numberOfTrees, withinSDtable,
  cladeAssignments, ID_names)
}
\arguments{
\item{distanceMatrix}{hk}

\item{numberOfTrees}{hkj}

\item{withinSDtable}{hkj}

\item{cladeAssignments}{hkj}

\item{ID_names}{hkj}
}
\description{
Bootstraps
}
