% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxillary_functions.R
\name{readAndCombineTrees}
\alias{readAndCombineTrees}
\title{Collates a folder of tree files into a single multiPhylo tree object.}
\usage{
readAndCombineTrees(pathToTrees, taxaToExclude = NULL, taxonToRoot = NULL)
}
\arguments{
\item{pathToTrees}{The path to your trees folder.}

\item{taxaToExclude}{Any taxa you would like excluded from the tree.  Useful to enforce conformity.}

\item{taxonToRoot}{Taxon to root all trees at.}
}
\value{
Output is a multiPhylo object.
}
\description{
Collates a folder of tree files into a single multiPhylo tree object.
}
