% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxillary_functions.R
\name{treesToCladeCompare}
\alias{treesToCladeCompare}
\title{Calculate in and out distances between and within clades for a set of trees.}
\usage{
treesToCladeCompare(multiPhyloTrees, cladeListFile = "list.txt",
  referenceTree = NULL, refCompareMethod = "edgeset")
}
\arguments{
\item{multiPhyloTrees}{The tree object.}

\item{cladeListFile}{The file.}

\item{referenceTree}{A reference tree.}

\item{refCompareMethod}{The method to use for dist.multiPhylo.}
}
\description{
Calculate in and out distances between and within clades for a set of trees.
}
