% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxillary_functions.R
\name{trimTreeFromList}
\alias{trimTreeFromList}
\title{Drops a set of tips from a tree.}
\usage{
trimTreeFromList(treeToTrim, listToDrop)
}
\arguments{
\item{treeToTrim}{The target tree that you would like to drop tips on.}

\item{listToDrop}{A list of tips to drop.}
}
\value{
Output is the trimmed tree of treeToTrim.
}
\description{
Drops a set of tips from a tree.
}
