% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxillary_functions.R
\name{trimTreeToTree}
\alias{trimTreeToTree}
\title{Trims the tips of a target tree A to match the tips of source tree B.}
\usage{
trimTreeToTree(treeToTrim, limitingTree)
}
\arguments{
\item{treeToTrim}{The target tree that you would like to drop tips on.}

\item{limitingTree}{The source tree that will determine which tips are dropped.}
}
\value{
Output is the trimmed tree of treeToTrim.
}
\description{
Trims the tips of a target tree A to match the tips of source tree B.
}
