% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_trees.R
\name{withinGroupDistances}
\alias{withinGroupDistances}
\title{Calculate the average distance between groups given a distance matrix and a clade}
\usage{
withinGroupDistances(distanceMatrix, cladeAssignments, ID_names)
}
\arguments{
\item{distanceMatrix}{The distance matrix}

\item{cladeAssignments}{The clade assignments}

\item{ID_names}{List of IDs}
}
\description{
Calculate the average distance between groups given a distance matrix and a clade
}
