% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_trees.R
\name{generateBootstrapTrees}
\alias{generateBootstrapTrees}
\title{Generates bootstraps trees based on within-group standard deviation.}
\usage{
generateBootstrapTrees(distanceMatrix, numberOfTrees = 1000, withinSDtable,
  cladeAssignments, ID_names)
}
\arguments{
\item{distanceMatrix}{A pairwise distance matrix.}

\item{numberOfTrees}{The number of bootstrap trees to generate. Default: 1000}

\item{withinSDtable}{The within clade SD table.}

\item{cladeAssignments}{Ordered list of clade assignments for taxa.}

\item{ID_names}{List of taxa names corresponding to the clade assignments.}
}
\description{
Generates bootstraps trees based on within-group standard deviation.
}
